backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.22.0",
    "google-genai>=1.18.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "pydantic>=2.5.1",
    "python-multipart>=0.0.20",
]
